The previous implementation of the "move completed files on removal" code was flawed, as it relied solely on DiskManager to do anything. After talking to parg, I've changed this so that the logic determining if the files should be moved (and where they should be moved to) is available as a static method on DiskManagerImpl.
The actual movement of the files can then be done by either the DiskManager or DownloadManager object. We no longer create a DiskManager to move the files, and we no longer perform the move inside GlobalManagerImpl (we've had to update a few places which will remove the DownloadManager globally to call downloadRemoved on the DownloadManager object).

As well as being more elegant, it's also appeared to fix the file locking bug in the original implementation.


git-svn-id: http://svn.vuze.com/public/client@8985 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2