MinimizedWindow is dead, long live the MiniBar!
First, the big structural change - the functionality which comprised the MinimizedWindow is now split up into different components in the new minibar package - loosely speaking, there's one class to manage all "mini bars", one class which provides a useful API for constructing "mini bars", and one class which uses this functionality to construct the download bar.

The MinimizedWindow class has been removed to prevent any further usage.

Secondly, the new functionality - we now have a new type of download bar, the "All Transfers" bar. Despite the fact I can't think of a good consistent name for it, and the fact it looks a little bit rubbish at the moment, we now have something which acts just like a download bar, except that it will contain the download and upload stats for all transfers - the right click menu will also provide a way to start / stop / pause / resume downloads. Plugins can register menu items on this bar too.

I think the "All Transfers" bar needs a bit of work on it, so I'm happy to take suggestions - but this very functional prototype will do for now.


git-svn-id: http://svn.vuze.com/public/client@10950 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2