Plugin menu items now appear in the Torrent menu. We have supported plugins being able to add to the torrent menu for a long time, but they have to explicitly register to do this.
Now what we do is take any table context menu items and fake a table row to appear in the Torrent menu.

What plugins should do is move to use DOWNLOAD_CONTEXT or TORRENT_MENU. In future, I'll remove this hack once more plugins move over to using the correct definition.

In the torrent menu, we differentiate between the two types of menu item, whereas we won't in the context menu.


git-svn-id: http://svn.vuze.com/public/client@12979 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2