Not happy with the change Gouss made - I reckon the default autoclose time for Mr Slidey was changed from 15 secs to 5 secs because of the "download was added" popup you get in the Vuze UI.
We might actually want to allow different values for popups that are caused by something the user has done (like adding a download), and those which they haven't (like a download completing).

Anyway - the Vuze UI now has a default of 5 secs, the Az UI has a default of 15 secs as before - that will have to do as a workaround.


git-svn-id: http://svn.vuze.com/public/client@18014 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2